Schedule generate 30 minutes after crawl
Observed that the pushes to the main branch from the GitHub Action by the bot were not triggering this GitHub Action to trigger a deployment.

While a solution to understand why this is happening would be ideal.  The fact this may be intentional to avoid infinite loops means that a scheduled deployment 30 minutes after a crawl is a reasonable workaround.

Looking at the repo there have been multiple crawls that have updated the file that in the logs did not trigger a redeployment of the pages.